fix(trpc): keep the page-comment unit tests out of the db import graph
page-comment.test.ts imported the router to reach agentSessionFor and
assertActivatedForAgent. The router pulls @superset/db/client, which calls
neon() at module scope, so the file threw on import in CI where DATABASE_URL
is unset — the same trap publish.integration.ts is named around.

Both helpers are pure, so they move to agent-access.ts beside the router,
matching page/access.ts and page/publish-rules.ts. No behaviour change.
